security_service/requirements.txt
alembic
SQLAlchemy
fastapi
uvicorn
psycopg2-binary
pydantic
PyJWT
httpx
loguru
pydantic-settings